requirements.txt
adrf==0.1.8
anyio==4.6.2.post1
asgiref==3.8.1
async-property==0.2.2
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.8
cryptography==43.0.3
distlib==0.3.9
Django==5.0.2
django-cors-headers==4.3.1
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==3.15.1
factory_boy==3.3.1
Faker==33.1.0
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
identify==2.6.3
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
MarkupSafe==3.0.2
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.0
pyOpenSSL==24.2.1
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==6.0.0
pytest-django==4.7.0
pytest-factoryboy==2.7.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
types-factory-boy==0.4.1
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==1.0.0
Werkzeug==3.1.3
setuptools==68.0.0
